+++ minigalaxy/installer.py
@@ -657,6 +657,13 @@ class InstallerInventory:
             file_info["md5"] = False
             return False
 
+
+    def get_expected_total_size(self):
+        total_size = 0
+        for file_data in self.data.values():
+            total_size += file_data.get('size', 0)
+        return total_size
+
     def is_complete(self):
         self.load()
         if not self.data: